+++ docs/source/.static/examples/plot_stations.py
@@ -3,11 +3,13 @@ Evaluate CHAOS at a Ground Observatory
 ======================================
 
 This script creates a time series plot of the first time-derivative of
-the magnetic field components (SV) from the CHAOS geomagnetic field model.
+the magnetic field components (SV) by evaluating the CHAOS geomagnetic field
+model.
 
 In this example the location of the ground observatory in Niemegk (Germany)
 is used. The spherical harmonic coefficients of the SV are truncated
-at degree 16.
+at degree 16. Note that the SV vector components are spherical geographic and
+not geodetic.
 
 """
 
@@ -17,14 +19,20 @@ import numpy as np
 
 model = cp.CHAOS.from_mat('CHAOS-7.mat')  # load the mat-file of CHAOS-7
 
+height = 0.  # geodetic height of the observatory in Niemegk (WGS84)
+lat_gg = 52.07  # geodetic latitude of Niemegk in degrees
+lon = 12.68  # longitude in degrees
+
+radius, theta = cp.coordinate_utils.gg_to_geo(height, 90. - lat_gg)
+
 data = {
     'Time': np.linspace(cp.mjd2000(1998, 1, 1), cp.mjd2000(2018, 1, 1), 500),  # time in mjd2000
-    'Radius': 6371.2,  # mean radius of Earth's surface in km
-    'Theta': 37.93,  # colatitude in degrees
-    'Phi': 12.68  # longitude in degrees
+    'Radius': radius,  # spherical geographic radius in kilometers
+    'Theta': theta,  # spherical geographic colatitude in degrees
+    'Phi': lon  # longitude in degrees
 }
 
-# compute SV components with CHAOS up to degree 16
+# compute SV components up to degree 16 (note deriv=1)
 dBr, dBt, dBp = model.synth_values_tdep(
     data['Time'], data['Radius'], data['Theta'], data['Phi'], nmax=16, deriv=1)