+++ CHANGELOG.rst
@@ -3,11 +3,14 @@ Changelog
 
 Version 0.13
 ------------
-| **Date:** November 4, 2023
+| **Date:** January 16, 2023
 | **Release:** v0.13
 
 Features
 ^^^^^^^^
+* Extended :func:`chaosmagpy.coordinate_utils.geo_to_gg` and
+  :func:`chaosmagpy.coordinate_utils.gg_to_geo` to also allow for the
+  rotation of vector components.
 * Removed Cartopy from the dependencies (basic map projections are done with
   Matplotlib and coastlines are loaded from Natural Earth). New dependency for
   reading the coastline shapefile is PyShp (>=2.3.1).